+++ test-fragments/org.ietr.preesm.evaluator.test/src/org/ietr/preesm/throughput/transformers/test/SDFTransformerTest.java
@@ -72,7 +72,7 @@ public class SDFTransformerTest {
     // check the normalization factor of each edge
     for (final SDFEdge e : sdf.edgeSet()) {
       final double Zt = (double) e.getSource().getPropertyBean().getValue("normalizedRate");
-      final double alpha_expected = Zt / e.getProd().intValue();
+      final double alpha_expected = Zt / e.getProd().longValue();
       final double alpha_current = (double) e.getPropertyBean().getValue("normalizationFactor");
       Assert.assertEquals(alpha_expected, alpha_current, 0);
     }
@@ -117,11 +117,11 @@ public class SDFTransformerTest {
 
     // verify that the consumption/production rate of all edges equal 1
     for (final SDFEdge e : hsdf.edgeSet()) {
-      final int cons = e.getCons().intValue();
-      final int prod = e.getProd().intValue();
+      final long cons = e.getCons().longValue();
+      final long prod = e.getProd().longValue();
 
-      Assert.assertEquals(1, cons);
-      Assert.assertEquals(1, prod);
+      Assert.assertEquals(1L, cons);
+      Assert.assertEquals(1L, prod);
     }
   }
 
@@ -145,11 +145,11 @@ public class SDFTransformerTest {
 
     // verify that the consumption/production rate of all edges equal 1
     for (final SDFEdge e : reducedHSDF.edgeSet()) {
-      final int cons = e.getCons().intValue();
-      final int prod = e.getProd().intValue();
+      final long cons = e.getCons().longValue();
+      final long prod = e.getProd().longValue();
 
-      Assert.assertEquals(1, cons);
-      Assert.assertEquals(1, prod);
+      Assert.assertEquals(1L, cons);
+      Assert.assertEquals(1L, prod);
     }
   }
 
@@ -182,7 +182,7 @@ public class SDFTransformerTest {
 
     // check if all the edges have zero delay
     for (final SDFEdge e : dag.edgeSet()) {
-      final int delay = e.getDelay().intValue();
+      final long delay = e.getDelay().longValue();
       Assert.assertEquals(0, delay);
 
     }
